+++ CHANGELOG.md
@@ -9,4 +9,8 @@ CHANGES:
   source have been renamed to `path`, `args` and `dir` to better reflect the
   CUE terminology.
 
+FEATURES:
+
+* Add attribute `pkg` to **cue_export** for the package to be loaded.
+
 ## 0.1.0 (Initial Release)