+++ CHANGELOG.md
@@ -5,9 +5,8 @@ CHANGES:
 * Provider has been migrated from `terraform-plugin-sdkv2` to
   `terraform-plugin-framework`.
 
-* Attributes `expression`, `files` and `working_dir` inside **cue_export** data
-  source have been renamed to `path`, `args` and `dir` to better reflect the
-  CUE terminology.
+* Attributes `expression` and `working_dir` inside **cue_export** data source has been renamed to
+  `path` and `dir`.
 
 FEATURES: