+++ CHANGELOG.md
@@ -8,6 +8,9 @@ CHANGES:
 * Attributes `expression`, `files` and `working_dir` inside **cue_export** data
   source have been renamed to `expr`, `paths` and `dir`.
 
+* The algorithm used for computing the `id` of the **cue_export** data source
+  has been changed from FNV-64a to FNV-128a.
+
 FEATURES:
 
 * Add attribute `pkg` to **cue_export** for the package to be loaded.